+++ apps/backend/src/bee-keeping/hive-inspection/hive-inspection.service.ts
@@ -175,9 +175,9 @@ export class HiveInspectionService {
     }
   }
 
-  async resetTable(): Promise<{ message: string }> {
+  async resetTable(userId: number): Promise<{ message: string }> {
     try {
-      await this.prisma.hive_inspection.deleteMany({});
+      await this.prisma.hive_inspection.deleteMany({ where: { bee_hives: { apiculture: { user_id: userId } } } });
       return { message: 'Hive Inspections table has been reset' };
     } catch (error) {
       throw new InternalServerErrorException(
