+++ apps/backend/src/bee-keeping/honey-production/honey-production.service.ts
@@ -123,9 +123,9 @@ export class HoneyProductionService {
     }
   }
 
-  async resetTable(): Promise<{ message: string }> {
+  async resetTable(userId: number): Promise<{ message: string }> {
     try {
-      await this.prisma.honey_production.deleteMany({});
+      await this.prisma.honey_production.deleteMany({ where: { bee_hives: { apiculture: { user_id: userId } } } });
       return { message: 'Honey Production table has been reset' };
     } catch {
       throw new InternalServerErrorException(
